improved tree generation, still needs some work

+++ VoxelEngine/src/Terrain/Chunk.hpp
@@ -16,7 +16,7 @@
 #include "TerrainConfig.hpp"
 #include "VoxelVertex.hpp"
 #include "Biome/Biome.hpp"
-#include "Generators/StructureGenerator.hpp"
+#include "Structures/StructureGenerator.hpp"
 
 namespace VoxelEngine
 {
@@ -66,7 +66,7 @@ private:
     void DetermineEdgeMeshes(VoxelMeshBuilder& meshBuilder, Voxel& v, size_t x, size_t z);
     void AddEdgeMesh(VoxelMeshBuilder& meshBuilder, Voxel& v, VoxelFace f);
     void AddEdgeMesh(VoxelMeshBuilder& meshBuilder, Voxel& v, VoxelFace f1, VoxelFace f2);
-    void DetermineVoxelFeatures(Voxel& v, size_t x, size_t z, size_t h) const;
+    void DetermineVoxelFeatures(Voxel& v, size_t x, size_t z, int32_t h) const;
     void AddStructures(const std::vector<Structure>& structures);
 
     void InitRadiance();
